make some refactor for modules.
