+++ controllers/authController.js
@@ -1,16 +1,16 @@
-const catchAsync = require("../../utils/catchAsync");
-const AppError = require("../../utils/appError");
-const Token = require("./Token");
+const catchAsync = require("../utils/catchAsync");
+const AppError = require("../errors/AppError");
+const Token = require("../services/Token");
 const { StatusCodes } = require("http-status-codes");
-const Sender = require("../../services/Sender");
-const Email = require("../../services/Email");
+const Sender = require("../services/Sender");
+const Email = require("../services/Email");
 const crypto = require("crypto");
-const db = require("../../models");
+const db = require("../models");
 const { Op } = require("sequelize");
-const stringToNumber = require("../../utils/stringToNumber");
-const authValidator = require("../../validators/authValidator");
-const validate = require("../../utils/validate");
-const Logger = require("../../services/Logger");
+const stringToNumber = require("../utils/stringToNumber");
+const authValidator = require("../validators/authValidator");
+const validate = require("../utils/validate");
+const Logger = require("../services/Logger");
 const authLogger = new Logger("auth");
 
 exports.loginAsParent = catchAsync(async (req, res, next) => {
