--- controllers/socketIOController/index.js
@@ -1,6 +1,0 @@
-module.exports = (io) => {
-  return {
-    protect: require("./protectMiddleware")(io),
-    connection: require("./onConnection")(io),
-  };
-};
