+++ controllers/subjectController.js
@@ -2,7 +2,7 @@ const db = require("../models");
 const factoryHandler = require("./factoryHandler");
 const subjectValidator = require("../validators/subjectValidator");
 const catchAsync = require("../utils/catchAsync");
-const AppError = require("../utils/appError");
+const AppError = require("../errors/AppError");
 const { StatusCodes } = require("http-status-codes");
 const Sender = require("../services/Sender");
 const { Op, literal } = require("sequelize");
