+++ errors/SocketError.js
