+++ routes/examRoutes/examRoutes.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const exams = require("../../controllers/examController");
 const restrictTo = require("../../utils/restrictTo");
-const protect = require("../../controllers/authController/protect");
+const protect = require("../../middlewares/protectMiddleware");
 const questions = require("./questionRoutes");
 const answers = require("./answerRoutes");
